Add package and context parameters to directives
