+++ talondoc/sphinx/directives/__init__.py
@@ -1,7 +1,13 @@
 from collections.abc import Callable
-from typing import TYPE_CHECKING, Any, Optional, cast
+from typing import TYPE_CHECKING, Any, Optional, Union, cast
 
 import sphinx.directives
+
+from talondoc.registry.entries.user import (
+    UserFileEntry,
+    UserPackageEntry,
+    UserTalonFileEntry,
+)
 
 if TYPE_CHECKING:
     from talondoc.sphinx.domains import TalonDomain
