+++ talondoc/sphinx/directives/command/__init__.py
@@ -1,327 +1,78 @@
-import re
 import sys
-from collections.abc import Callable, Iterator
-from typing import Optional, Union
+from typing import List, Optional
 
-from docutils import nodes
 from sphinx import addnodes
 from sphinx.util.typing import OptionSpec
 from talonfmt import talonfmt
-from tree_sitter_talon import (
-    TalonCapture,
-    TalonChoice,
-    TalonComment,
-    TalonEndAnchor,
-    TalonList,
-    TalonOptional,
-    TalonParenthesizedRule,
-    TalonRepeat,
-    TalonRepeat1,
-    TalonRule,
-    TalonSeq,
-    TalonStartAnchor,
-    TalonWord,
-)
+from typing_extensions import override
 
-from ....registry import Registry
-from ....registry.entries.user import (
-    UserCommandEntry,
-    UserPackageEntry,
-    UserTalonFileEntry,
+from talondoc.sphinx.directives.command.util import (
+    describe_command,
+    find_commands,
+    include_command,
 )
-from ....sphinx.directives import TalonDocDirective, TalonDocObjectDescription
-from ....sphinx.typing import TalonDocstringHook_Callable
-from ....util.desc import InvalidInterpolation
-from ....util.describer import TalonScriptDescriber
+from talondoc.sphinx.directives.errors import AmbiguousSignature, UnmatchedSignature
+
+from ....registry.entries.user import UserCommandEntry
+from ....sphinx.directives import TalonDocObjectDescription
 from ....util.logging import getLogger
-from ....util.nodes import desc_content, desc_name, paragraph
-from ....util.typing import flag
+from ....util.typing import flag, optional_strlist
 
 _LOGGER = getLogger(__name__)
-
-
-_RE_WHITESPACE = re.compile(r"\s+")
-
-
-def include_command(
-    candidate: UserCommandEntry,
-    sig: Optional[str] = None,
-    *,
-    fullmatch: bool = True,
-    default: str = "include",
-    include: tuple[str, ...] = (),
-    exclude: tuple[str, ...] = (),
-    captures: Optional[
-        Callable[
-            [str],
-            Optional[
-                Union[
-                    TalonCapture,
-                    TalonChoice,
-                    TalonEndAnchor,
-                    TalonList,
-                    TalonOptional,
-                    TalonParenthesizedRule,
-                    TalonRepeat,
-                    TalonRepeat1,
-                    TalonRule,
-                    TalonSeq,
-                    TalonStartAnchor,
-                    TalonWord,
-                ]
-            ],
-        ]
-    ] = None,
-    lists: Optional[Callable[[str], Optional[list[str]]]] = None,
-):
-    assert default in ["include", "exclude"]
-
-    def match(sig: str) -> bool:
-        try:
-            words = _RE_WHITESPACE.split(sig)
-            return bool(
-                candidate.ast.rule.match(
-                    words,
-                    fullmatch=fullmatch,
-                    get_capture=captures,
-                    get_list=lists,
-                )
-            )
-        except IndexError:
-            return False
-
-    def excluded() -> bool:
-        return (
-            bool(exclude)
-            and any(match(exclude_sig) for exclude_sig in exclude)
-            and not any(match(include_sig) for include_sig in include)
-        )
-
-    def included() -> bool:
-        return any(match(include_sig) for include_sig in include) and not any(
-            match(exclude_sig) for exclude_sig in exclude
-        )
-
-    if default == "include":
-        return (not sig or match(sig)) and not excluded()
-    else:
-        return (sig and match(sig)) or included()
-
-
-def describe_rule(command: UserCommandEntry) -> nodes.Text:
-    return nodes.Text(talonfmt(command.ast.left, safe=False))
-
-
-def try_describe_script_via_action_docstrings(
-    command: UserCommandEntry,
-    *,
-    registry: Registry,
-    docstring_hook: Optional[TalonDocstringHook_Callable],
-) -> Optional[nodes.Text]:
-    """
-    Describe the script using the docstrings on the actions used by the script.
-    """
-    try:
-        describer = TalonScriptDescriber(registry, docstring_hook=docstring_hook)
-        desc = describer.describe(command.ast)
-        if desc:
-            return nodes.Text(str(desc))
-    except InvalidInterpolation as e:
-        _LOGGER.exception(e)
-    return None
-
-
-def try_describe_script_via_script_docstrings(
-    command: UserCommandEntry,
-) -> Optional[nodes.Text]:
-    """
-    Describe the script using the docstrings present in the script itself.
-    """
-    comments = []
-    for child in command.ast.right.children:
-        if isinstance(child, TalonComment):
-            comment = child.text.strip()
-            if comment.startswith("###"):
-                comments.append(comment.removeprefix("###").strip())
-    if comments:
-        return nodes.Text(" ".join(comments))
-    else:
-        return None
-
-
-def describe_script_via_code_block(command: UserCommandEntry) -> addnodes.highlightlang:
-    """
-    Describe the script by including it as a code block.
-    """
-    code = talonfmt(command.ast.script, safe=False)
-    literal_block = nodes.literal_block(code, code)
-    literal_block["language"] = "python"
-    return literal_block  # type: ignore
-
-
-def describe_script(
-    command: UserCommandEntry,
-    *,
-    registry: Registry,
-    include_script: bool,
-    docstring_hook: Optional[TalonDocstringHook_Callable],
-) -> list[nodes.Element]:
-    """
-    Describe the script using the docstrings on the script, the docstrings on
-    the actions, or finally by including it as a code block.
-    """
-    buffer = []
-
-    # 1. Use the Talon docstrings in the command script. Talon docstrings are comments which start with ###.
-    desc = try_describe_script_via_script_docstrings(command)
-
-    # 2. Use the Python docstrings from the actions used in the command script.
-    if desc is None:
-        desc = try_describe_script_via_action_docstrings(
-            command, registry=registry, docstring_hook=docstring_hook
-        )
-
-    # Append the description.
-    if desc is not None:
-        buffer.append(paragraph(nodes.Text(str(desc))))
-
-    if desc is None or include_script:
-        # 3. Include the command script.
-        buffer.append(describe_script_via_code_block(command))
-
-    return buffer
-
-
-def describe_command(
-    command: UserCommandEntry,
-    signode: addnodes.desc_signature,
-    *,
-    registry: Registry,
-    include_script: bool,
-    docstring_hook: Optional[TalonDocstringHook_Callable],
-) -> addnodes.desc_signature:
-    signode += desc_name(describe_rule(command))
-    signode += desc_content(
-        *describe_script(
-            command,
-            registry=registry,
-            include_script=include_script,
-            docstring_hook=docstring_hook,
-        )
-    )
-    return signode
 
 
 class TalonCommandDirective(TalonDocObjectDescription):
     has_content = True
     required_arguments = 1
     optional_arguments = sys.maxsize
-    option_spec: OptionSpec = {"script": flag}
+    option_spec: OptionSpec = {
+        "package": optional_strlist,
+        "context": optional_strlist,
+        "include_script": flag,
+    }
     final_argument_whitespace = False
 
-    def get_signatures(self):
-        return [" ".join(self.arguments)]
+    @override
+    def get_signatures(self) -> List[str]:
+        sig = " ".join(self.arguments)
+        return [sig]
+
+    @override
+    def handle_signature(self, sig: str, signode: addnodes.desc_signature):
+        try:
+            command = self.find_command(sig)
+            include_script = self.options.get("include_script", False)
+            describe_command(
+                command,
+                signode,
+                registry=self.talon.registry,
+                include_script=include_script,
+                docstring_hook=self.docstring_hook,
+            )
+            return command.get_name()
+        except (UnmatchedSignature, AmbiguousSignature) as e:
+            _LOGGER.error(e)
+            raise ValueError(e)
 
     def find_command(self, sig: str) -> UserCommandEntry:
-        command: Optional[UserCommandEntry] = None
-        for candidate in self.talon.registry.commands:
-            if include_command(candidate, sig, fullmatch=True):
-                if __debug__ and command:
-                    raise ValueError(f"Signature '{sig}' matched multiple commands.")
-                command = candidate
-                if not __debug__:
-                    break
-        if command:
-            return command
-        else:
-            raise ValueError(f"Signature '{sig}' matched no commands.")
-
-    def handle_signature(self, sig: str, signode: addnodes.desc_signature):
-        command = self.find_command(sig)
-        describe_command(
-            command,
-            signode,
-            registry=self.talon.registry,
-            include_script=self.options.get("script", False),
-            docstring_hook=self.docstring_hook,
-        )
-        return command.get_name()
-
-
-class TalonCommandListDirective(TalonDocDirective):
-    def find_package(self) -> UserPackageEntry:
-        namespace = self.options.get("package")
-        candidate = self.talon.registry.active_package_entry
-        if candidate and (not namespace or candidate.get_namespace() == namespace):
-            return candidate
-        candidate = self.talon.registry.packages.get(namespace, None)
-        if candidate:
-            return candidate
-        raise ValueError(f"Could not find package '{namespace}'")
-
-    def find_file(
-        self, sig: str, *, package_entry: Optional[UserPackageEntry] = None
-    ) -> UserTalonFileEntry:
-        # Try lookup with 'sig' as name:
-        result = self.talon.registry.lookup("file", sig)
-        if result:
-            return result
-
-        # Try lookup with 'sig.talon' as name:
-        result = self.talon.registry.lookup("file", f"{sig}.talon")
-        if result:
-            return result
-
-        # Try searching in the active package:
-        if package_entry is None:
-            try:
-                package_entry = self.find_package()
-            except ValueError as e:
-                raise ValueError(f"Could not find file '{sig}'", e)
-
-        # Find the file:
-        for file in package_entry.files:
-            if isinstance(file, UserTalonFileEntry):
-                # Try comparison with 'sig' as path:
-                if sig == str(file.path):
-                    return file
-                # Try comparison with 'sig.talon' as path:
-                if f"{sig}.talon" == str(file.path):
-                    return file
-        raise ValueError(f"Could not find file '{sig}'.")
-
-    def find_commands(self) -> Iterator[UserCommandEntry]:
-        exclude = self.options.get("exclude", ())
-        include = self.options.get("include", ())
-
-        if self.arguments:
-            # If file arguments were given, return all commands in that file
-            # which have not been explicitly excluded:
-            for sig in self.arguments:
-                default = self.options.get("default", "include")
-                for command in self.find_file(sig).commands:
-                    if include_command(
-                        command, default=default, exclude=exclude, include=include
-                    ):
-                        yield command
-        else:
-            # If no file arguments were given, return all commands in the package
-            # which have been explicitly included:
-            default = self.options.get("default", "exclude")
-            for command in self.talon.registry.commands:
-                if include_command(
-                    command, default=default, exclude=exclude, include=include
-                ):
-                    yield command
-
-    def caption(self) -> Iterator[str]:
-        # Get caption from options
-        caption = self.options.get("caption", None)
-        if caption:
-            yield caption
-            return
-        # Get caption from file name
-        if len(self.arguments) == 1:
-            file = self.find_file(self.arguments[0])
-            yield file.get_name().removesuffix(".talon")
-            return
+        packages = self.options.get("package", [])
+        contexts = self.options.get("context", [])
+        commands = [
+            command
+            for command in find_commands(
+                self.talon.registry, contexts=contexts, packages=packages
+            )
+            if include_command(command, sig, fullmatch=True)
+        ]
+        if len(commands) == 0:
+            raise UnmatchedSignature(self.get_location(), sig)
+        if len(commands) > 1:
+            raise AmbiguousSignature(
+                self.get_location(),
+                sig,
+                [
+                    f"{command.get_location()}: {talonfmt(command.ast.left, safe=False)}"
+                    for command in commands
+                ],
+            )
+        return commands[0]
